requirements.txt
flask~=3.0.3
ultralytics~=8.2.79
opencv-python~=4.10.0.84
torch~=2.1.0
python-dotenv~=1.0.1
